refactor: split 3 god-modules into focused sub-modules
Extract handler and stage functions from server-handlers.js (1087 LOC),
pre-loop-stages.js (910 LOC), and iteration-stages.js (891 LOC) into
smaller focused modules under handlers/ and stages/ directories.

Original files become thin re-exporters preserving the public API.
All 2473 tests pass with zero changes to functionality.
